# Documentation/openapi.yaml
openapi: 3.0.3
info:
  title: NetBanking Microservices Platform API
  description: |
    Comprehensive single-file OpenAPI 3.0 specification for the NetBanking Distributed Banking Platform.
    This specification aggregates all API surfaces across the 5 core microservices:
    1. **Authentication Service** (`auth-service` - Port 8081)
    2. **User & Customer Profile Service** (`user-service` - Port 8082)
    3. **Account & Ledger Service** (`account-service` - Port 8083)
    4. **Transaction & Statement Service** (`transaction-service` - Port 8084)
    5. **Notification & Audit Service** (`notification-service` - Port 8085)
    
    All external traffic routes through the Spring Cloud API Gateway (Port 8080) with pure JWT Bearer authentication.
  version: 1.0.0
  contact:
    name: NetBanking Platform Engineering
    email: support@netbanking.com
servers:
  - url: http://localhost:8080
    description: Spring Cloud API Gateway (Primary Edge Gateway)
  - url: http://localhost:8081
    description: Direct Auth Service
  - url: http://localhost:8082
    description: Direct User Service
  - url: http://localhost:8083
    description: Direct Account Service
  - url: http://localhost:8084
    description: Direct Transaction Service
  - url: http://localhost:8085
    description: Direct Notification Service
...
paths:
...
  /api/v1/auth/register:
    post:
      tags:
        - Authentication & Identity
      summary: Register a new customer identity
      description: Validates email uniqueness, creates AUTH_USER credentials with BCrypt hashing, initializes the customer shell in User-Service, and issues an email verification OTP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration successful, verification OTP issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid input or email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/login:
    post:
      tags:
        - Authentication & Identity
      summary: Authenticate user credentials and issue pure JWTs
      description: Validates email and password. If MFA is enabled and OTP is omitted, generates a 6-digit login OTP and returns mfaRequired=true. Once verified, issues HMAC-SHA256 signed access and refresh tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful or MFA OTP challenge initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Account locked or inactive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/refresh:
    post:
      tags:
        - Authentication & Identity
      summary: Refresh an expired access token
      description: Validates an unrevoked refresh token, revokes it (rotation), and issues a fresh pure JWT access token and new refresh token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Successfully refreshed tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid, expired, or revoked refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/logout:
    post:
      tags:
        - Authentication & Identity
      summary: Revoke session and logout
      description: Revokes the refresh token and registers a LOGOUT event in the login history.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: Session terminated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/v1/auth/verify-otp:
    post:
      tags:
        - Authentication & Identity
      summary: Verify one-time password (OTP)
      description: Verifies 6-digit OTP code for EMAIL_VERIFICATION, LOGIN MFA, or PASSWORD_RESET purposes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOtpRequest'
      responses:
        '200':
          description: OTP verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/forgot-password:
    post:
      tags:
        - Authentication & Identity
      summary: Request password reset instructions
      description: Generates a secure reset token and issues a password reset OTP/link via email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Reset instructions dispatched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/v1/auth/reset-password:
    post:
      tags:
        - Authentication & Identity
      summary: Reset password with token
      description: Consumes a valid password reset token and updates user password hash.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # =========================================================================
  # 2. USER & CUSTOMER PROFILE SERVICE ENDPOINTS
  # =========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Provide the pure JWT access token issued by Auth Service (`Bearer <token>`).
  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: customer@netbanking.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 100
          example: SecureP@ssw0rd!
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: customer@netbanking.com
        password:
          type: string
          format: password
          example: SecureP@ssw0rd!
        otp:
          type: string
          description: 6-digit OTP code (if MFA is enabled and challenging)
          example: "123456"
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: Pure HMAC-SHA256 signed JWT
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          description: Refresh token string
          example: 8e5d0f19-bca1-4567-8910-111213141516
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          format: int64
          description: Access token lifetime in seconds
          example: 900
        customerId:
          type: string
          example: C94A82F12B4C
        roles:
          type: array
          items:
            type: string
          example: ["CUSTOMER"]
        mfaRequired:
          type: boolean
          example: false
    RefreshRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    LogoutRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    VerifyOtpRequest:
      type: object
      required:
        - email
        - otp
        - purpose
      properties:
        email:
          type: string
          format: email
        otp:
          type: string
          pattern: '^\\d{6}$'
          example: "748291"
        purpose:
          type: string
          enum: [LOGIN, PASSWORD_RESET, EMAIL_VERIFICATION]
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
        newPassword:
          type: string
          minLength: 8
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: Operation completed successfully.
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Email is already registered
...